editoast: isolate infra CLI into its own module
